+++ src/bzflag/HUDRenderer.cxx
@@ -1365,7 +1365,7 @@ void			HUDRenderer::renderPlaying(SceneRenderer& renderer)
   LocalPlayer *myTank = LocalPlayer::getMyTank();
   if (myTank && myTank->getPosition()[2] < 0.0f) {
     glColor4f(0.02f, 0.01f, 0.01f, 1.0);
-    glRectf(0, 0, width, (myTank->getPosition()[2]/(BZDB->eval(StateDatabase::BZDB_BURROWDEPTH)-0.1f)) * (viewHeight/2.0f));
+    glRectf(0, 0, (float)width, (myTank->getPosition()[2]/(BZDB->eval(StateDatabase::BZDB_BURROWDEPTH)-0.1f)) * ((float)viewHeight/2.0f));
   }
